requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Web scraping and data extraction
beautifulsoup4>=4.12.0
Pillow>=10.0.0

# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Slack API
slack-sdk>=3.23.0

# AI/ML for categorization and summarization
openai>=1.3.0

# Web framework for API
flask>=3.0.0
flask-cors>=4.0.0

# Template engine
jinja2>=3.1.2

# Data handling
pyyaml>=6.0.1
